move assert_big_text_equal to nbconvert
where it's used
