+++ IPython/testing/tools.py
@@ -465,23 +465,3 @@ def help_all_output_test(subcommand=''):
     nt.assert_in("Class parameters", out)
     return out, err
 
-def assert_big_text_equal(a, b, chunk_size=80):
-    """assert that large strings are equal
-
-    Zooms in on first chunk that differs,
-    to give better info than vanilla assertEqual for large text blobs.
-    """
-    for i in range(0, len(a), chunk_size):
-        chunk_a = a[i:i + chunk_size]
-        chunk_b = b[i:i + chunk_size]
-        nt.assert_equal(chunk_a, chunk_b, "[offset: %i]\n%r != \n%r" % (
-            i, chunk_a, chunk_b))
-
-    if len(a) > len(b):
-        nt.fail("Length doesn't match (%i > %i). Extra text:\n%r" % (
-            len(a), len(b), a[len(b):]
-        ))
-    elif len(a) < len(b):
-        nt.fail("Length doesn't match (%i < %i). Extra text:\n%r" % (
-            len(a), len(b), b[len(a):]
-        ))
