+++ jupyter_nbconvert/exporters/tests/test_notebook.py
@@ -9,7 +9,7 @@ from .base import ExportersTestsBase
 from ..notebook import NotebookExporter
 
 from IPython.nbformat import validate
-from IPython.testing.tools import assert_big_text_equal
+from jupyter_nbconvert.tests.base import assert_big_text_equal
 
 class TestNotebookExporter(ExportersTestsBase):
     """Contains test functions for notebook.py"""
